# backend/docs/swagger.yaml
basePath: /api
host: edis.dshub.com.ng
info:
  contact:
    email: support@lifegate.health
    name: LifeGate Support
  description: AI-powered clinical diagnostic platform. Patients interact with an
    AI assistant, physicians review AI-generated reports, and administrators manage
    the full lifecycle.
  license:
    name: Proprietary
  termsOfService: https://lifegate.health/terms
  title: LifeGate API
  version: "1.0"
paths:
...
  /alerts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  alerts:
                    type: array
                  total:
                    type: integer
                type: object
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Patient alerts
      tags:
      - alerts
...
  /physician/alerts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  alerts:
                    type: array
                  total:
                    type: integer
                type: object
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Physician alerts
      tags:
      - alerts
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"